🎨 Formatted files with Prettier and moved types to lib/
