+++ .github/workflows/stale.yml
@@ -2,14 +2,14 @@ name: Stale
 
 on:
   schedule:
-  - cron: "30 1 * * *"
+    - cron: "30 1 * * *"
 
 jobs:
   stale:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/stale@v3
-      with:
-        repo-token: ${{ secrets.GITHUB_TOKEN }}
-        stale-issue-message: 'This issue has been automatically marked as stale because it has not had recent activity 😢 It will be closed if no further activity occurs. Thank you for your contributions.'
-        stale-pr-message: 'This PR has been automatically marked as stale because it has not had recent activity 😭 It will be closed if no further activity occurs. Thank you for your contributions.'
+      - uses: actions/stale@v3
+        with:
+          repo-token: ${{ secrets.GITHUB_TOKEN }}
+          stale-issue-message: "This issue has been automatically marked as stale because it has not had recent activity 😢 It will be closed if no further activity occurs. Thank you for your contributions."
+          stale-pr-message: "This PR has been automatically marked as stale because it has not had recent activity 😭 It will be closed if no further activity occurs. Thank you for your contributions."
