+++ src/backend/events/ipc/core.ts
@@ -1,7 +1,7 @@
 import tabEvents from "./tab";
 import windowEvents from "./window";
 import Window from "../../models/Window";
-import { HandlerProps } from "../../types/index";
+import { HandlerProps } from "../../lib/types";
 
 const handler = (window: Window, props: HandlerProps) => {
   windowEvents(window);
